# api/docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.CreatePickingList:
    properties:
      coming_increment_id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  models.CreateProduct:
    properties:
      branch_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  models.PickingList:
    properties:
      coming_id:
        type: string
      coming_increment_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      total_price:
        type: number
      updated_at:
        type: string
    type: object
  models.Product:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  models.UpdatePickingList:
    properties:
      coming_id:
        type: string
      coming_increment_id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  models.UpdateProduct:
    properties:
      branch_id:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /picking_list:
    get:
      consumes:
      - application/json
      description: Get List PickingList details by its ok.
      parameters:
      - description: Offset
        in: path
        name: offset
        type: integer
      - description: Limit
        in: path
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: PickingList details
          schema:
            $ref: '#/definitions/models.PickingList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: PickingList not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get List PickingList
      tags:
      - PickingList
    post:
      consumes:
      - application/json
      description: Create PickingList
      parameters:
      - description: PickingList ID
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.CreatePickingList'
      produces:
      - application/json
      responses:
        "200":
          description: PickingList details
          schema:
            $ref: '#/definitions/models.PickingList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: PickingList not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: create a PickingList
      tags:
      - PickingList
  /picking_list/{id}:
    delete:
      consumes:
      - application/json
      description: Get List PickingList details by its ok
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PickingList details
          schema:
            $ref: '#/definitions/models.PickingList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: PickingList not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get List PickingList
      tags:
      - PickingList
    get:
      consumes:
      - application/json
      description: Get PickingList details by its ID.
      parameters:
      - description: PickingList ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PickingList details
          schema:
            $ref: '#/definitions/models.PickingList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: PickingList not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a PickingList by ID
      tags:
      - PickingList
    put:
      consumes:
      - application/json
      description: Get List PickingList details by its ok.
      parameters:
      - description: models.UpdatePickingList
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePickingList'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PickingList details
          schema:
            $ref: '#/definitions/models.PickingList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: PickingList not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update PickingList
      tags:
      - PickingList
  /product:
    get:
      consumes:
      - application/json
      description: Get List Product details by its ok.
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get List Product
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Create Product
      parameters:
      - description: Product ID
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.CreateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: create a Product
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Get List Product details by its ok
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get List Product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get Product details by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a Product by ID
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Get List Product details by its ok.
      parameters:
      - description: models.UpdateProduct
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProduct'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Product
      tags:
      - Product
...
swagger: "2.0"